+++ src/main/java/io/github/tr100000/codec2schema/api/Codec2SchemaPlugin.java
@@ -1,0 +1,7 @@
+package io.github.tr100000.codec2schema.api;
+
+public interface Codec2SchemaPlugin {
+    default void earlyRegisterHandlers() {}
+    default void registerHandlers() {}
+    default void generateSchemas(SchemaExporter exporter) {}
+}