+++ src/main/java/io/github/tr100000/codec2schema/api/CodecHandlerEarlyRegistrationEntrypoint.java
@@ -1,6 +1,7 @@
 package io.github.tr100000.codec2schema.api;
 
 @FunctionalInterface
+@Deprecated(forRemoval = true)
 public interface CodecHandlerEarlyRegistrationEntrypoint {
     void earlyRegister();
 }