+++ src/main/java/io/github/tr100000/codec2schema/api/SchemaGenerationEntrypoint.java
@@ -1,6 +1,7 @@
 package io.github.tr100000.codec2schema.api;
 
 @FunctionalInterface
+@Deprecated(forRemoval = true)
 public interface SchemaGenerationEntrypoint {
     void generate(SchemaExporter exporter);
 }