+++ CHANGELOG.md
@@ -1,3 +1,5 @@
-### Fixed
-- Fixed erroneous entries in required fields in some cases
-- Fixed crash without Fabric API (hopefully)
+### Changed
+- Replaced entrypoints with `codec2schema:main` and `codec2schema:client`, these must implement `Codec2SchemaPlugin`
+  - Old entrypoints will no longer work, this is a **breaking change**
+- Split client and common sources
+  - This will help prevent bugs and crashes in the future